Исправил ошибки
1) Добавил поток ThreadS2VNA в котором организована работа с socket , принимает и обрабатывает данные
2) Исправил работу с коллиеством частотных точек,  отрисовка адекватная
3) Перенес часть кода из main.cpp в MainWindow.cpp
4) При изменении размера окна, размеры графика тоже изменяются
5) Сбор данных происходит в циклическом режиме в классе ThreadS2VNA

+++ scpicommands.cpp
@@ -18,13 +18,13 @@ QString SCPICommands::SetFilterBand(const int &channel, const int &filterBand)
 }
 
 // Устанавлваем начальную частоту
-QString SCPICommands::SetStartFrequency(const int &channel, const float &startFrequency)
+QString SCPICommands::SetStartFrequency(const int &channel, const double &startFrequency)
 {
     return "SENS" + QString::number(channel) + ":FREQ:STAR " + QString::number(startFrequency) + "GHZ\n";
 }
 
 // Установка конечной частоты
-QString SCPICommands::SetStopFrequency(const int &channel, const float &stopFrequency)
+QString SCPICommands::SetStopFrequency(const int &channel, const double &stopFrequency)
 {
     return "SENS" + QString::number(channel) + ":FREQ:STOP " + QString::number(stopFrequency) + "GHZ\n";
 }
@@ -34,6 +34,8 @@ QString SCPICommands::SelectionMeasuredParameter(const int &channel, const int &
 {
     return "CALC" + QString::number(channel) + ":PAR" + QString::number(chart) + ":DEF " + parameter + "\n";
 }
+
+
